requirements.txt
# Core RAG and AI/ML dependencies
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Financial data and analysis
yfinance>=0.2.0
yahoofinancials>=0.8.9
alpha-vantage>=2.3.1
fredapi>=0.5.0
ta-lib>=0.4.25
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
newspaper3k>=0.2.8
feedparser>=6.0.10

# News and sentiment analysis
newsapi-python>=0.2.6
textblob>=0.17.1
vaderSentiment>=3.3.2
nltk>=3.8.1

# Time series analysis
statsmodels>=0.14.0
arch>=6.2.0
prophet>=1.1.4

# Web framework and API
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database and storage
sqlalchemy>=2.0.0
sqlite3
redis>=4.6.0

# Utilities and helpers
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Data processing and validation
pydantic>=2.0.0
marshmallow>=3.20.0
jsonschema>=4.19.0

# Security and authentication
cryptography>=41.0.0
passlib>=1.7.4

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Optional: GPU support
# cupy-cuda11x>=12.0.0  # Uncomment for CUDA support
